if game already exists dont create it again
