+++ frontend/src/components/ArtistGrid.js
@@ -1,51 +1,5 @@
-import * as React from "react";
-import { useState, useEffect } from "react";
-import { ReactComponent as NextBtn } from "../images/next_btn.svg";
-import StaticDataService from "../services/StaticDataService";
-import { fetchSession } from "../utils";
-import UserService from "../services/UserService";
-
 export const ArtistGrid = (props) => {
-  const { selectedId, handleClick } = props;
-  const [artistData, setArtistData] = useState([]);
-  const [artistImages, setArtistImages] = useState({});
-  const [games, setGames] = useState([]);
-
-  const fetchArtist = async (id) => {
-    try {
-      const artist = await StaticDataService.getArtists();
-      return artist;
-    } catch (err) {
-      console.log(err, "Error fetching the artist");
-    }
-  };
-
-  const fetchGamesByUserId = async () => {
-    try {
-      const user = await UserService.getGameByUserId(
-        fetchSession().id, //userid
-        "",
-        fetchSession().token
-      );
-
-      return user;
-    } catch (err) {
-      console.log(err, "Error fetching user");
-    }
-  };
-
-  useEffect(() => {
-    const init = async () => {
-      const artists = await fetchArtist();
-      let gamesArray = await fetchGamesByUserId();
-      const images = await StaticDataService.getArtistImages();
-      setArtistImages(images);
-      setGames(gamesArray);
-      setArtistData(artists);
-    };
-
-    init();
-  }, []);
+  const { selectedId, handleClick, artistData, artistImages, games } = props;
 
   const renderButtons = (startHere, endHere) => {
     if (artistData.length > 0) {
